Issue #18 Upadted name of class according to the ADD
